s2 - server. styles refactoring
