+++ server/index.js
@@ -7,7 +7,7 @@ import mongoose from "mongoose";
 import filesRouter from "./routers/files-router.js"; */
 import groupRouter from "./routers/group-router.js";
 import memberRouter from "./routers/member-router.js";
-/* import styleRouter from "./routers/style-router.js"; */
+import styleRouter from "./routers/style-router.js";
 
 
 const PORT = 5000;
@@ -21,8 +21,9 @@ app.use(cors());
 
 app.use("/api", groupRouter);
 app.use("/api", memberRouter);
-/* app.use("/api", albumRouter);
 app.use("/api", styleRouter);
+/* app.use("/api", albumRouter);
+
 app.use("/api", filesRouter); */
 
 async function startApp() {
