+++ server/services/styles-service.js
@@ -1,0 +1,6 @@
+import Style from "../models/style.js";
+import Services from "./services.js";
+
+class StylesService extends Services {}
+
+export default new StylesService(Style);
